refactor(lib): extract shared adapter primitives into toolbox-store-common.ts
Dedupes the transport + identity-extraction boilerplate between
src/lib/toolbox-store.ts (mentions: hn/reddit/bluesky/devto) and
src/lib/toolbox-store-velocity.ts (stars + fork velocity).

Extracted to src/lib/toolbox-store-common.ts:
- DEFAULT_LIMIT, DEFAULT_TIMEOUT_MS constants
- ToolboxEvent, ToolboxLeaderboardResponse, ToolboxFetchOptions types
- extractGithubFullName (was identical in both files)
- fetchLeaderboardEvents (was identical in both files)
- maxProducedAtMs (was identical in both files)

Both adapters now import from common; behavior unchanged. Each adapter
still owns its own row-shaping logic (event → legacy file shape).

This unblocks B.9 (5 more mention adapters: huggingface×3, producthunt,
openai-rss) — they'll import the shared primitives instead of copying
the ~80 lines of boilerplate per adapter.

## Tests

All 30 tests pass (21 toolbox-store + 9 toolbox-store-velocity):
$ npx tsx --test --require ./tests/setup-server-only-stub.cjs \
    src/lib/__tests__/toolbox-store.test.ts \
    src/lib/__tests__/toolbox-store-velocity.test.ts

Net diff: ~80 lines moved to common; both adapter files lose duplicated
helpers but gain a clean import block.
