+++ src/lib/toolbox-store-velocity.ts
@@ -28,79 +28,14 @@ import "server-only";
 
 import type { DeltasJson, DeltaValue, RepoDeltaEntry } from "./trending";
 import { getFullNameToRepoId } from "./trending";
+import type { ToolboxEvent, ToolboxFetchOptions } from "./toolbox-store-common";
+import {
+  extractGithubFullName,
+  fetchLeaderboardEvents,
+  maxProducedAtMs,
+} from "./toolbox-store-common";
 
-const DEFAULT_LIMIT = 500;
-const DEFAULT_TIMEOUT_MS = 8_000;
-
-interface ToolboxEvent {
-  target_id: string;
-  target_kind: string;
-  target_identity: string;
-  scan_id: string;
-  run_id: string;
-  signal_type: string;
-  produced_at: string;
-  fields: Record<string, unknown>;
-}
-
-interface ToolboxLeaderboardResponse {
-  count: number;
-  targets: ToolboxEvent[];
-}
-
-export interface ToolboxFetchOptions {
-  apiUrl: string;
-  apiKey: string;
-  limit?: number;
-  timeoutMs?: number;
-  fetchImpl?: typeof fetch;
-}
-
-function extractGithubFullName(targetIdentity: string): string | null {
-  try {
-    const url = new URL(targetIdentity);
-    const host = url.hostname.toLowerCase();
-    if (host !== "github.com" && host !== "www.github.com") return null;
-    const parts = url.pathname.split("/").filter(Boolean);
-    if (parts.length < 2) return null;
-    return `${parts[0]}/${parts[1]}`;
-  } catch {
-    return null;
-  }
-}
-
-async function fetchLeaderboardEvents(
-  opts: ToolboxFetchOptions,
-  signalType: string,
-): Promise<ToolboxLeaderboardResponse | null> {
-  const limit = opts.limit ?? DEFAULT_LIMIT;
-  const timeoutMs = opts.timeoutMs ?? DEFAULT_TIMEOUT_MS;
-  const fetchImpl = opts.fetchImpl ?? globalThis.fetch;
-  if (typeof fetchImpl !== "function") return null;
-
-  const base = opts.apiUrl.replace(/\/+$/, "");
-  const url = `${base}/v1/signals/leaderboard?type=${encodeURIComponent(signalType)}&limit=${limit}`;
-
-  const ac = new AbortController();
-  const timer = setTimeout(() => ac.abort(), timeoutMs);
-  try {
-    const res = await fetchImpl(url, {
-      headers: {
-        authorization: `Bearer ${opts.apiKey}`,
-        accept: "application/json",
-      },
-      signal: ac.signal,
-    });
-    if (!res.ok) return null;
-    const body = (await res.json()) as ToolboxLeaderboardResponse;
-    if (!body || !Array.isArray(body.targets)) return null;
-    return body;
-  } catch {
-    return null;
-  } finally {
-    clearTimeout(timer);
-  }
-}
+export type { ToolboxFetchOptions };
 
 const WINDOWS = ["1h", "24h", "7d", "30d"] as const;
 type DeltaWindow = (typeof WINDOWS)[number];
@@ -150,15 +85,6 @@ function createEmptyEntry(): RepoDeltaEntry {
     delta_7d: noHistory(),
     delta_30d: noHistory(),
   };
-}
-
-function maxProducedAtMs(events: ToolboxEvent[]): number {
-  let max = 0;
-  for (const ev of events) {
-    const t = Date.parse(ev.produced_at);
-    if (Number.isFinite(t) && t > max) max = t;
-  }
-  return max;
 }
 
 /**
