+++ src/lib/toolbox-store.ts
@@ -43,46 +43,14 @@ import type {
   RedditPost,
   RedditRepoMention,
 } from "./reddit";
+import type { ToolboxEvent, ToolboxFetchOptions } from "./toolbox-store-common";
+import {
+  extractGithubFullName,
+  fetchLeaderboardEvents,
+  maxProducedAtMs,
+} from "./toolbox-store-common";
 
-const DEFAULT_LIMIT = 500;
-const DEFAULT_TIMEOUT_MS = 8_000;
-
-interface ToolboxEvent {
-  target_id: string;
-  target_kind: string;
-  target_identity: string;
-  scan_id: string;
-  run_id: string;
-  signal_type: string;
-  produced_at: string;
-  fields: Record<string, unknown>;
-}
-
-interface ToolboxLeaderboardResponse {
-  count: number;
-  targets: ToolboxEvent[];
-}
-
-export interface ToolboxFetchOptions {
-  apiUrl: string;
-  apiKey: string;
-  limit?: number;
-  timeoutMs?: number;
-  fetchImpl?: typeof fetch;
-}
-
-function extractGithubFullName(targetIdentity: string): string | null {
-  try {
-    const url = new URL(targetIdentity);
-    const host = url.hostname.toLowerCase();
-    if (host !== "github.com" && host !== "www.github.com") return null;
-    const parts = url.pathname.split("/").filter(Boolean);
-    if (parts.length < 2) return null;
-    return `${parts[0]}/${parts[1]}`;
-  } catch {
-    return null;
-  }
-}
+export type { ToolboxFetchOptions };
 
 function slugIdFromFullName(fullName: string): string {
   return String(fullName)
@@ -90,48 +58,6 @@ function slugIdFromFullName(fullName: string): string {
     .replace(/\//g, "--")
     .replace(/\./g, "-")
     .replace(/[^a-z0-9-]/g, "");
-}
-
-async function fetchLeaderboardEvents(
-  opts: ToolboxFetchOptions,
-  signalType: string,
-): Promise<ToolboxLeaderboardResponse | null> {
-  const limit = opts.limit ?? DEFAULT_LIMIT;
-  const timeoutMs = opts.timeoutMs ?? DEFAULT_TIMEOUT_MS;
-  const fetchImpl = opts.fetchImpl ?? globalThis.fetch;
-  if (typeof fetchImpl !== "function") return null;
-
-  const base = opts.apiUrl.replace(/\/+$/, "");
-  const url = `${base}/v1/signals/leaderboard?type=${encodeURIComponent(signalType)}&limit=${limit}`;
-
-  const ac = new AbortController();
-  const timer = setTimeout(() => ac.abort(), timeoutMs);
-  try {
-    const res = await fetchImpl(url, {
-      headers: {
-        authorization: `Bearer ${opts.apiKey}`,
-        accept: "application/json",
-      },
-      signal: ac.signal,
-    });
-    if (!res.ok) return null;
-    const body = (await res.json()) as ToolboxLeaderboardResponse;
-    if (!body || !Array.isArray(body.targets)) return null;
-    return body;
-  } catch {
-    return null;
-  } finally {
-    clearTimeout(timer);
-  }
-}
-
-function maxProducedAtMs(events: ToolboxEvent[]): number {
-  let max = 0;
-  for (const ev of events) {
-    const t = Date.parse(ev.produced_at);
-    if (Number.isFinite(t) && t > max) max = t;
-  }
-  return max;
 }
 
 function isoFromMaxOrNow(maxMs: number): string {
